Move singleton/logger.ts to logger dir
