+++ src/logger/index.ts
@@ -1,1 +1,2 @@
+export * from './logger'
 export * from './LogLevel'
