--- src/logger/singleton/index.ts
@@ -1,1 +1,0 @@
-export * from './logger'
